requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
jinja2==3.1.4
sqlalchemy==2.0.36
python-multipart==0.0.12
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==2.10.3
pydantic-settings==2.6.1
torf==4.3.0
guessit==3.8.0
